requirements.txt
PyYAML==6.0
requests==2.28.2
tabulate==0.9.0
